fix formatting and compilation issue due to dfg structs
