+++ crates/environ/src/component/translate/inline.rs
@@ -45,13 +45,14 @@
 //! side-effectful initializers are emitted to the `GlobalInitializer` list in the
 //! final `Component`.
 
-use crate::component::dfg::{CoreInstanceStructure, RuntimeComponentInstanceStructure, Source};
+pub use self::info::*;
 use crate::component::translate::*;
 use crate::{EntityType, IndexType};
 use core::str::FromStr;
 use sha2::{Digest, Sha256};
 use std::borrow::Cow;
 use std::ops::Index;
+use wasmparser::collections::Map;
 use wasmparser::component_types::{ComponentAnyTypeId, ComponentCoreModuleTypeId};
 
 pub(super) fn run(
@@ -377,7 +378,7 @@ fn record_core_def_from_export(
     name: &str,
     def: ComponentItemDef,
     types: &ComponentTypesBuilder,
-    map: &mut HashMap<String, dfg::CoreDef>,
+    map: &mut Map<String, dfg::CoreDef>,
 ) -> Result<()> {
     match &def {
         ComponentItemDef::Instance(instance) => match instance {
@@ -491,24 +492,28 @@ impl<'a> Inliner<'a> {
         }
     }
 
-    /// Records the runtime sources for each component export, 
-    /// so the instantiation graph later knows which 
+    /// Records the runtime sources for each component export,
+    /// so the instantiation graph later knows which
     /// core definitions they depend on.
-    /// Here internal component exports are recorded in addition to the 
-    /// exports recoreded for the top level component.
+    /// Here internal component exports are recorded in addition to the
+    /// exports recorded for the top level component.
     fn add_component_exports(
         &mut self,
         types: &mut ComponentTypesBuilder,
         exports: &IndexMap<&str, ComponentItemDef>,
         fr: &mut InlinerFrame,
     ) -> Result<()> {
-        let mut comp_exports: HashMap<String, dfg::CoreDef> = HashMap::new();
+        let mut comp_exports: Map<String, dfg::CoreDef> = Map::new();
         for (name, item) in exports.iter() {
             record_core_def_from_export(name, item.clone(), types, &mut comp_exports)?;
         }
         let mut count = 0;
         for (_, func) in comp_exports {
-            let source = self.core_def_to_sources.get(&func).unwrap_or(&Source::host()).clone();
+            let source = self
+                .core_def_to_sources
+                .get(&func)
+                .unwrap_or(&Source::host())
+                .clone();
             fr.instance_structure
                 .component_exports
                 .insert(count, source);
@@ -1309,8 +1314,8 @@ impl<'a> Inliner<'a> {
             // and an initializer is recorded to indicate that it's being
             // instantiated.
             ModuleInstantiate(module, args) => {
-                let mut core_imports: HashMap<u32, String> = Default::default();
-                let mut sources: HashMap<u32, Source> = Default::default();
+                let mut core_imports: Map<u32, String> = Default::default();
+                let mut sources: Map<u32, Source> = Default::default();
 
                 let (instance_module, init) = match &frame.modules[*module] {
                     ModuleDef::Static(idx, _ty) => {
@@ -1382,7 +1387,7 @@ impl<'a> Inliner<'a> {
                 match &frame.modules[*module] {
                     ModuleDef::Static(midx, _ty) => {
                         let core_instance = frame.module_instances.len() as u32;
-                        let mut core_exports: HashMap<u32, String> = HashMap::new();
+                        let mut core_exports: Map<u32, String> = Map::new();
                         let mut count = 0;
                         for (name, &entity) in self.nested_modules[*midx].module.exports.iter() {
                             core_exports.insert(count, name.to_string());
