+++ crates/wasmtime/src/runtime/component/component.rs
@@ -18,7 +18,7 @@ use std::path::Path;
 use wasmtime_environ::component::{
     CompiledComponentInfo, ComponentArtifacts, ComponentTypes, CoreDef, Export, ExportIndex,
     GlobalInitializer, InstantiateModule, NameMapNoIntern, OptionsIndex, StaticModuleIndex,
-    TrampolineIndex, TypeComponentIndex, TypeFuncIndex, UnsafeIntrinsic, VMComponentOffsets, dfg,
+    TrampolineIndex, TypeComponentIndex, TypeFuncIndex, UnsafeIntrinsic, VMComponentOffsets, info,
 };
 use wasmtime_environ::{Abi, CompiledFunctionsTable, FuncKey, TypeTrace};
 use wasmtime_environ::{FunctionLoc, HostPtr, ObjectKind, PrimaryMap};
@@ -882,7 +882,7 @@ impl Component {
     }
 
     /// Returns the Instantiation Graph.
-    pub fn instantiation_graph(&self) -> &dfg::RootComponentInstanceStructure {
+    pub fn instantiation_graph(&self) -> &info::RootComponentInstanceStructure {
         &self.inner.info.component.instantiation_graph
     }
 }
