+++ crates/wasmtime/src/runtime/component/mod.rs
@@ -131,9 +131,9 @@ pub use self::func::{
 };
 pub use self::has_data::*;
 pub use self::instance::{Instance, InstanceExportLookup, InstancePre};
-pub use self::linker::{Linker, LinkerInstance};
 #[cfg(feature = "caller")]
 pub use self::linker::CallerInstance;
+pub use self::linker::{Linker, LinkerInstance};
 pub use self::resource_table::{ResourceTable, ResourceTableError};
 pub use self::resources::{Resource, ResourceAny, ResourceDynamic};
 pub use self::types::{ResourceType, Type};
